refactor shopping list page with ~~vibes~~
